test(misc): rename phase3 test files to descriptive names
Splits the UI coverage tests into admin_updates_test.go (updates
tab, system checks, admin helpers) and sidebar_fragments_test.go
(worker-detail, deployment-log, search-users, tag assign). Renames
apps_phase3_test.go to apps_test.go — the existing api_test.go
covers the full surface, leaving no prior owner for that filename.
"Phase 3" was ticket bookkeeping with no meaning in the codebase.
